server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Atanas Pozharliev <ABPozharliev19@codingburgas.bg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.3"}
aerich = "^0.7.2"


[tool.aerich]
tortoise_orm = "data.data.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
